+++ src/style.ts
@@ -7,22 +7,20 @@ export const panelStyle = `
   .${config.panelClassName}{
     --border-color: #dee0e3;
     border: 1px solid var(--border-color);
-    left: 0;top:0;
+    left: 0; top: 0;
     background-color: #fff;
     display: none;
     width: ${config.defaultMenuWidth}px;
-    position:absolute;
+    position: absolute;
     z-index: ${config.defaultZIndex};
   }
 `;
 
 export const contextMenuStyle = `
   .${config.wrapperClassName}{
-    -webkit-user-select:none;
     user-select: none;
     padding: 2px 0 2px 0px;
     margin: 0;
-    cursor: default;
   }
   /*子菜单*/
   .${config.wrapperClassName} .${config.panelClassName}{ 
@@ -38,30 +36,26 @@ export const contextMenuStyle = `
     padding: 0 30px 0 30px;
     list-style: none;
     line-height: ${config.menuItemHeight}px;
-    font-size: 13px;
+    font-size: 12px;
     display: flex;
     justify-content: space-between;
-    flex-wrap: nowrap;
   }
   .${config.wrapperClassName} li.disabled{
     color: #5f6368;
     pointer-events: none;
   }
   .${config.wrapperClassName} li .menu-item-icon{
-    position:absolute;
-    left: 7px;
-    top: 4px;
+    position: absolute;
+    left: 7px; top: 4px;
+    width: 16px;
     height: 16px;
-    vertical-align: middle;
   }
   .${config.wrapperClassName} li .menu-item-label {
     overflow: hidden;
     text-overflow: ellipsis;
-    font-size: 12px;
   }
   .${config.wrapperClassName} li .menu-item-tip {
     color: #5f6368;
-    font-size: 12px;
   }
   .${config.wrapperClassName} li:hover:not(.divide):not(.disabled),
   .${config.wrapperClassName} li.${config.wrapperClassName}_hover {
@@ -76,13 +70,10 @@ export const contextMenuStyle = `
   }
   .${config.wrapperClassName} li .right-arrow {
     position: absolute;
-    right: 8px;
-    top: 9px;
+    right: 8px; top: 9px;
     border-top: 4px solid transparent;
-    border-left: 4px solid #000;
     border-right: 4px solid transparent;
     border-bottom: 4px solid transparent;
-  }
-  .${config.wrapperClassName}_child{
+    border-left: 4px solid #000;
   }
   `;